update-algolia netlify function setup to push proper Algolia data
