+++ website/functions/utils/return-response.js
@@ -1,0 +1,16 @@
+async function returnResponse(status, res) {
+  const headers = {
+    'Content-Type': 'application/json',
+    'Access-Control-Allow-Origin': '*',
+    'Access-Control-Allow-Headers': '*',
+    'Access-Control-Allow-Methods': 'POST, OPTIONS' 
+  }
+  const resObj = {
+    statusCode: status,
+    headers,
+    body: JSON.stringify(res)
+  }
+  return resObj
+}
+
+module.exports = returnResponse
